+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 0.8.1 - 2017-02-23
+
+* Dependency update.
+
+* Fixed deprecation warnings from `image-webpack-loader`. See https://github.com/tcoopman/image-webpack-loader/issues/68
+* Fixed deprecation warnings from `webpack 2`. 
+  * `fallbackLoader` option - Replaced with `fallback`.
+  * `loader` option - Replaced with `use`.
+
 ## 0.8.0 - 2017-01-18
 
 * Replaced unmaintained `isparta` with `nyc` and `babel-plugin-istanbul`.